(PC-30721)[API] feat: rework GDPR extraction to work with a cron instead of a cloud task
