+++ api/.env.development
@@ -123,3 +123,5 @@ DEMARCHES_SIMPLIFIEES_BANK_ACCOUNT_PROCEDURE_ID=80876
 ENABLE_TEST_ROUTES=1
 TITELIVE_EPAGINE_API_URL=https://catsearch.epagine.fr/v1
 TITELIVE_EPAGINE_API_AUTH_URL=https://login.epagine.fr/v1
+GDPR_MAX_EXTRACT_PER_DAY=10
+GDPR_LOCK_TIMEOUT=900
