+++ api/.env.integration
@@ -92,3 +92,5 @@ ZENDESK_SELL_BACKEND=pcapi.core.external.zendesk_sell_backends.logger.LoggerBack
 ENABLE_TEST_ROUTES=0
 TITELIVE_EPAGINE_API_URL=https://catsearch.epagine.fr/v1
 TITELIVE_EPAGINE_API_AUTH_URL=https://login.epagine.fr/v1
+GDPR_MAX_EXTRACT_PER_DAY=10
+GDPR_LOCK_TIMEOUT=900
