+++ api/.env.ops
@@ -122,3 +122,5 @@ SESSION_COOKIE_SAMESITE="None"
 ENABLE_TEST_ROUTES=1
 TITELIVE_EPAGINE_API_URL=https://catsearch.epagine.fr/v1
 TITELIVE_EPAGINE_API_AUTH_URL=https://login.epagine.fr/v1
+GDPR_MAX_EXTRACT_PER_DAY=10
+GDPR_LOCK_TIMEOUT=900
