+++ api/.env.staging
@@ -118,3 +118,5 @@ BIG_QUERY_TABLE_BASENAME="passculture-data-ehp.backend_stg"
 ENABLE_TEST_ROUTES=1
 TITELIVE_EPAGINE_API_URL=https://catsearch.epagine.fr/v1
 TITELIVE_EPAGINE_API_AUTH_URL=https://login.epagine.fr/v1
+GDPR_MAX_EXTRACT_PER_DAY=10
+GDPR_LOCK_TIMEOUT=900
