+++ api/.env.testauto
@@ -58,3 +58,5 @@ BOOST_API_USERNAME=fake_user
 BOOST_API_PASSWORD=fake_password
 GCP_GDPR_EXTRACT_BUCKET=gdpr-bucket
 GCP_GDPR_EXTRACT_FOLDER=extracts
+GDPR_MAX_EXTRACT_PER_DAY=10
+GDPR_LOCK_TIMEOUT=900
