+++ api/src/pcapi/tasks/__init__.py
@@ -9,6 +9,5 @@ def install_handlers(app: Flask) -> None:
     from . import beamer_tasks
     from . import external_api_booking_notification_tasks
     from . import finance_tasks
-    from . import gdpr_tasks
     from . import sendinblue_tasks
     from . import ubble_tasks
